+++ Utopia/Utopia/Worlds/Chunks/WorldChunks.Update.cs
@@ -55,7 +55,7 @@ namespace Utopia.Worlds.Chunks
             }
         }
 
-        public override void VTSUpdate(double interpolationHd, float interpolationLd, long elapsedTime)
+        public override void VTSUpdate(double interpolationHd, float interpolationLd, float elapsedTime)
         {
             ChunkVisibilityTest();
